refactor : Controller, Service 패키지 변경 및 URI 수정
