+++ src/main/java/wooteco/subway/service/PathService.java
@@ -1,4 +1,4 @@
-package wooteco.subway.service.client;
+package wooteco.subway.service;
 
 import java.util.List;
 import java.util.stream.Collectors;
